rename yt-theater component and move it to it's own file
