+++ assembl/static2/js/app/components/debate/common/post/postBody.jsx
@@ -1,11 +1,11 @@
 // @flow
 import React from 'react';
-import get from 'lodash/get';
 import activeHtml from 'react-active-html';
 
 import PostTranslate from '../../common/translations/postTranslate';
 import { transformLinksInHtml } from '../../../../utils/linkify';
 import { youtubeRegexp } from '../../../../utils/globalFunctions';
+import YoutubeTheater from '../../../common/youtubeTheater';
 
 type Props = {
   body: string,
@@ -20,62 +20,13 @@ type Props = {
   translationEnabled: boolean
 };
 
-class YouTubeWithTheater extends React.Component {
-  state: {
-    theaterMode?: boolean
-  };
-
-  timeout: number;
-
-  closeTheater = () => this.setState({ theaterMode: false });
-
-  openTheater = () => this.setState({ theaterMode: true });
-
-  render = () => {
-    const { videoId } = this.props;
-    const theaterMode = get(this, 'state.theaterMode', false);
-    const video = (
-      <iframe
-        title="YouTube video"
-        id="ytplayer"
-        type="text/html"
-        width="640"
-        height="360"
-        src={`https://www.youtube.com/embed/${videoId}?autoplay=1`}
-        frameBorder="0"
-      />
-    );
-    return (
-      <div className={`youtube-theater ${theaterMode ? 'open' : ''}`}>
-        {theaterMode ? (
-          <div className="theater-content">
-            <div className="youtube-video">
-              {video}
-              {theaterMode && <button onClick={this.closeTheater} className="close-theater-button assembl-icon-cancel" />}
-            </div>
-          </div>
-        ) : (
-          <div
-            className="youtube-thumbnail-container"
-            onClick={this.openTheater}
-            style={{ backgroundImage: `https://img.youtube.com/vi/${videoId}/mqdefault.jpg` }}
-          >
-            <img className="youtube-thumbnail" alt="youtube video" src={`https://img.youtube.com/vi/${videoId}/mqdefault.jpg`} />
-            <span className="play-button" />
-          </div>
-        )}
-      </div>
-    );
-  };
-}
-
 const postBodyReplacementComponents = {
   iframe: (attributes) => {
     const { src } = attributes;
     const regexpMatch = src.match(youtubeRegexp);
     if (regexpMatch) {
       const videoId = regexpMatch[1];
-      return <YouTubeWithTheater videoId={videoId} />;
+      return <YoutubeTheater videoId={videoId} />;
     }
     return <iframe title="post-embed" {...attributes} />;
   }
